글 수정 기능 오류 수정 시작

+++ application/views/todo/view_contents_v.php
@@ -54,7 +54,12 @@
             <tr>
                 <th colspan="6">
                     <a href="http://34.80.199.17/Main/lists/" class="btn btn-primary">목록</a>
-                    <a href="http://34.80.199.17/Main/delete/<?php echo $this -> uri -> segment(3); ?>/<?php echo $transID; ?>/<?php echo $transAccountID; ?>" class="btn btn-danger">삭제</a>
+                    <?php if ($accountInfo == $transAccountID || $permit->permit == 1):?>
+                        <a href="http://34.80.199.17/Main/delete/<?php echo $this -> uri -> segment(3); ?>/<?php echo $transID; ?>/<?php echo $transAccountID; ?>" class="btn btn-danger">삭제</a>
+                    <?php else:?>
+                        <a href="http://34.80.199.17/Main/lists/" class="btn btn-danger" onclick="alert('자신이 작성한 글만 삭제 가능합니다.')">삭제불가</a>
+                    <?php endif;?>
+
                     <a href="http://34.80.199.17/Main/write/" class="btn btn-success">쓰기</a>
                     <?php if($accountInfo == $transAccountID || $permit->permit == 1):?>
                             <a href="http://34.80.199.17/Main/edit/<?php echo $this -> uri -> segment(3);?>" class="btn btn-success">수정</a>
